requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
kafka-python==2.0.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
pyspark==3.5.0
boto3==1.34.0
google-generativeai==0.3.2
elevenlabs==0.2.26
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
pydantic-settings==2.1.0
